refactor: split landing page into modular components
Addresses code review suggestions for better maintainability:

- Externalize data: Move all data arrays to src/data/landing.ts and src/config/contact.ts
- Split components: Extract 8 focused sections (Hero, Rooms, Dining, Experiences, Technology, Operations, Awards, Contact)
- Add error boundaries: Wrap lazy-loaded ConciergeChatWidget with ErrorBoundary component
- Improve loading states: Add spinner and message for lazy-loaded components
- Extract custom hooks: usePrefersReducedMotion and useRevealOnScroll for better reusability
- Reduce page.tsx from 622 to 154 lines (75% reduction)

Benefits:
- Improved code organization and maintainability
- Better performance through proper code splitting
- Enhanced error handling for lazy-loaded components
- Easier to test individual components
- Configurable contact information

All builds pass successfully. Lint errors are pre-existing in unrelated legacy scripts.
