+++ src/components/ErrorBoundary.tsx
@@ -1,0 +1,46 @@
+'use client'
+
+import { Component, type ReactNode } from 'react'
+
+interface Props {
+  children: ReactNode
+  fallback?: ReactNode
+}
+
+interface State {
+  hasError: boolean
+  error?: Error
+}
+
+export default class ErrorBoundary extends Component<Props, State> {
+  constructor(props: Props) {
+    super(props)
+    this.state = { hasError: false }
+  }
+
+  static getDerivedStateFromError(error: Error): State {
+    return { hasError: true, error }
+  }
+
+  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
+    console.error('ErrorBoundary caught an error:', error, errorInfo)
+  }
+
+  render() {
+    if (this.state.hasError) {
+      if (this.props.fallback) {
+        return this.props.fallback
+      }
+
+      return (
+        <div className="rounded-2xl border border-lundies-stone/60 bg-lundies-linen/50 p-6 text-center">
+          <p className="text-sm text-lundies-peat">
+            Something went wrong loading this component. Please refresh the page.
+          </p>
+        </div>
+      )
+    }
+
+    return this.props.children
+  }
+}
