+++ src/config/contact.ts
@@ -1,0 +1,5 @@
+export const contactChannels = [
+  { label: 'Visit', detail: '1-5 Dunkeld Street, Aberfeldy, PH15 2AA', href: 'https://maps.app.goo.gl/qxgTB1gYkn12' },
+  { label: 'Reserve', detail: '+44 (0)1887 352 214', href: 'tel:+441887352214' },
+  { label: 'Enquire', detail: 'hello@schiehallionhotel.co.uk', href: 'mailto:hello@schiehallionhotel.co.uk' },
+]
