+++ src/data/landing.ts
@@ -1,0 +1,141 @@
+export const landingSections = [
+  { label: 'Story', href: '#story' },
+  { label: 'Rooms', href: '#rooms' },
+  { label: 'Dining', href: '#dining' },
+  { label: 'Experiences', href: '#experiences' },
+  { label: 'Technology', href: '#technology' },
+  { label: 'Operations', href: '#operations' },
+  { label: 'Awards', href: '#awards' },
+  { label: 'Contact', href: '#contact' },
+]
+
+export const heroImages = [
+  {
+    src: '/images/rooms/generated/suite/suite-artistic-1759248824246-ur9015.png',
+    alt: 'Soft morning light inside the Schiehallion loft suite',
+    caption: 'Loch Tay suites with tailored service',
+  },
+  {
+    src: '/images/rooms/generated/deluxe/deluxe-artistic-1759248801864-outd2c.png',
+    alt: 'Deluxe room featuring freestanding bath and herringbone floors',
+    caption: 'Deluxe comforts with Highland textures',
+  },
+  {
+    src: '/images/rooms/generated/family/family-artistic-1759248840189-b78zgy.png',
+    alt: 'Family suite with warm timber panelling and seating nook',
+    caption: 'Family suites crafted for together time',
+  },
+]
+
+export const roomCollections = [
+  {
+    name: 'Highland View Double',
+    summary: 'Loch-side calm, super-king sleep, fresh pastries on cue.',
+    image: '/images/rooms/generated/standard/standard-artistic-1759248767421-r0mln6.png',
+    meta: 'Signature Double · 2 adults + child',
+  },
+  {
+    name: 'Riverside Twin Retreat',
+    summary: 'Twin comfort with ergonomic workspace and ultrafast Wi-Fi.',
+    image: '/images/rooms/generated/accessible/accessible-artistic-1759248857443-57441r.png',
+    meta: 'Twin · Work-ready sanctuary',
+  },
+  {
+    name: 'Family Cairngorm Suite',
+    summary: 'Interconnected rooms, games chest, and concierge planning.',
+    image: '/images/rooms/generated/family/family-artistic-1759248840189-b78zgy.png',
+    meta: 'Two-bedroom suite · 2 adults · 2 children',
+  },
+  {
+    name: 'Deluxe Schiehallion Loft',
+    summary: 'Freestanding bath, private check-in, telescope over Ben Lawers.',
+    image: '/images/rooms/generated/deluxe/deluxe-artistic-1759248801864-outd2c.png',
+    meta: 'Loft suite · Celebrations',
+  },
+]
+
+export const diningMoments = [
+  {
+    title: 'The Schiehallion Kitchen',
+    description: 'Seasonal tasting journeys with Perthshire produce and sommelier pairings.',
+    image: '/images/rooms/generated/suite/suite-artistic-1759248824246-ur9015.png',
+    highlights: ['Four-course tasting', 'Vegan pairings', 'Wine flights'],
+  },
+  {
+    title: 'Breakfast & Brunch',
+    description: 'Sunrise spreads, Glen Lyon coffee, and take-away baskets for river guides.',
+    image: '/images/rooms/generated/standard/standard-artistic-1759248767421-r0mln6.png',
+    highlights: ['Fresh pastries', 'Express takeaway', 'Locally roasted beans'],
+  },
+  {
+    title: 'Afternoon Rituals',
+    description: 'Highland shortbread towers, live roast countdown, patisserie made in-house.',
+    image: '/images/rooms/generated/accessible/accessible-artistic-1759248857443-57441r.png',
+    highlights: ['Seasonal tea menu', 'Allergen aware', 'Sunday roast theatre'],
+  },
+]
+
+export const experienceStories = [
+  {
+    name: 'Dewars Aberfeldy Distillery',
+    distance: '2 miles',
+    description: 'Heritage tastings, storytelling lounges, bespoke whisky flights.',
+  },
+  {
+    name: 'Beyond Adventure',
+    distance: 'On the River Tay',
+    description: 'Kayaking, paddleboarding, and private guides for every level.',
+  },
+  {
+    name: 'Loch Tay & Ben Lawers',
+    distance: '15 minutes',
+    description: 'Sunrise hikes, photographer-led expeditions, snowline picnics.',
+  },
+  {
+    name: 'Aberfeldy Watermill & Gallery',
+    distance: '5 minute walk',
+    description: 'Independent bookshop, art exhibitions, slow-travel inspiration.',
+  },
+]
+
+export const technologyPillars = [
+  {
+    title: 'Unified Booking Surface',
+    detail:
+      'Next.js 15 powering responsive web, PWA, and admin journeys with shared tokens and inclusive patterns.',
+  },
+  {
+    title: 'Real-time Availability Core',
+    detail: 'Firestore + Realtime Database syncing room calendars, dining tables, and partner allotments.',
+  },
+  {
+    title: 'Edge-first Performance',
+    detail: 'CDN image optimisation, caching strategies, and sub-100ms first content worldwide.',
+  },
+  {
+    title: 'Payments & Assurance',
+    detail: 'Stripe integrations with 3D Secure, automated confirmations, and audit-ready histories.',
+  },
+]
+
+export const operationsHighlights = [
+  {
+    title: 'Operations Command Centre',
+    points: ['Drag-and-drop room assignment', 'Housekeeping snapshots', 'Proactive maintenance alerts'],
+  },
+  {
+    title: 'Revenue Studio',
+    points: ['Dynamic pricing experiments', 'Package performance analytics', 'Competitive benchmarks'],
+  },
+  {
+    title: 'Guest Relationship Desk',
+    points: ['Unified inbox', 'Pre-arrival and post-stay flows', 'Loyalty segmentation'],
+  },
+]
+
+export const awards = [
+  { label: 'Scottish Hotel Awards', detail: 'Boutique Hotel of the Year · 2023' },
+  { label: 'Green Key Certification', detail: 'Sustainable operations accreditation' },
+  { label: 'VisitScotland', detail: '5-Star Gold Guest Accommodation' },
+  { label: 'Condé Nast Traveller', detail: 'Readers Choice Top 20 Highlands' },
+]
